Check if a BlockRef contains a nil block

+++ src/server/pfs/server/obj_block_api_server.go
@@ -357,7 +357,7 @@ func (s *objBlockAPIServer) DeleteObjects(ctx context.Context, request *pfsclien
 				return err
 			}
 
-			if objectInfo != nil && objectInfo.BlockRef != nil {
+			if objectInfo != nil && objectInfo.BlockRef != nil && objectInfo.BlockRef.Block != nil {
 				blockPath := s.localServer.blockPath(objectInfo.BlockRef.Block)
 				if err := s.objClient.Delete(blockPath); err != nil && !s.isNotFoundErr(err) {
 					return err
